requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Django==5.1.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==8.0.33
protobuf==3.20.3
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
WTForms==3.2.1
